correction

+++ src/main/java/com/xuanzhe_franck/openfood/Repository/NutritionScoreRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NutritionScoreRepository extends JpaRepository<NutritionScore, Long> {
+public interface NutritionScoreRepository extends JpaRepository<NutritionScore, Integer> {
 
 }
